+++ common/src/main/java/team/tnt/collectorsalbum/common/card/CardUiTemplate.java
@@ -7,21 +7,24 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import team.tnt.collectorsalbum.common.AlbumCategory;
 import team.tnt.collectorsalbum.common.init.SoundRegistry;
 import team.tnt.collectorsalbum.platform.Codecs;
 
+import java.util.Optional;
 import java.util.function.Supplier;
 
-public record CardUiTemplate(Integer[] effectColors, Integer[] effectDurations, Supplier<SoundEvent> flipSoundRef) {
+public record CardUiTemplate(Integer[] effectColors, Integer[] effectDurations, Supplier<SoundEvent> flipSoundRef, ResourceLocation cardTexture) {
 
-    public static final CardUiTemplate TEMPLATE = new CardUiTemplate(new Integer[] { 0x00FF00 }, new Integer[] { 5 }, SoundRegistry.FLIP_COMMON);
+    public static final CardUiTemplate TEMPLATE = new CardUiTemplate(new Integer[] { 0x00FF00 }, new Integer[] { 5 }, SoundRegistry.FLIP_COMMON, null);
     public static final Codec<CardUiTemplate> CODEC = RecordCodecBuilder.<CardUiTemplate>create(instance -> instance.group(
             Codecs.array(Codec.INT, Integer[]::new).optionalFieldOf("effectColors", TEMPLATE.effectColors()).forGetter(t -> t.effectColors),
             Codecs.array(Codec.INT, Integer[]::new).optionalFieldOf("effectDurations", TEMPLATE.effectDurations()).forGetter(t -> t.effectDurations),
-            Codecs.supplier(BuiltInRegistries.SOUND_EVENT.byNameCodec()).optionalFieldOf("flipSound", TEMPLATE.flipSoundRef()).forGetter(t -> t.flipSoundRef)
-    ).apply(instance, CardUiTemplate::new)).validate(template -> template.effectColors().length == template.effectDurations().length
+            Codecs.supplier(BuiltInRegistries.SOUND_EVENT.byNameCodec()).optionalFieldOf("flipSound", TEMPLATE.flipSoundRef()).forGetter(t -> t.flipSoundRef),
+            ResourceLocation.CODEC.optionalFieldOf("cardTexture").forGetter(t -> Optional.ofNullable(t.cardTexture))
+    ).apply(instance, (colors, durations, sound, textureOpt) -> new CardUiTemplate(colors, durations, sound, textureOpt.orElse(null)))).validate(template -> template.effectColors().length == template.effectDurations().length
             ? DataResult.success(template)
             : DataResult.error(() -> "Effect duration count has to be the same as effect colors")
     );