+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-# 2.1.5
-- Fixed card duplication when using shift-click to extract cards from album #27
-- Fixed server side disconnect when sending datapacks to all connected clients #28
+# 2.1.6
+- Added `blacklist` property to entity drop filter - [Documentation](https://github.com/T-T-Team/CollectorsAlbum/wiki/Item-drops#entity-filter)
+- Added `cardTexture` field to cards to allow you to customize texture used in card pack opening UI - [Documentation](https://github.com/T-T-Team/CollectorsAlbum/wiki/Cards)